==> 5_identify_key_subj_obj_pred_from_text_complete.ipynb <==
import pandas as pd
import numpy as np
import re
import warnings
warnings.filterwarnings("ignore")
import math
import os
import datetime
from pandas import DataFrame
import itertools
from bs4 import BeautifulSoup
import urllib.request
import requests
import random
import torch
import xmltodict
import xml.etree.ElementTree as ET
import xml
from xml.etree.ElementTree import parse
from xml.parsers import expat
from xml.dom.minidom import parse, parseString
import psycopg2
import PyPDF2
import pikepdf
from bs4 import BeautifulSoup
import nltk
nltk.download('punkt')
nltk.download('averaged_perceptron_tagger')
import re
import spacy

def createfiledirectory(directory):
    try:
        if not os.path.exists(directory):
            os.makedirs(directory)
    except OSError:
        print ('Error: Creating directory. ' +  directory)
    return directory

# **LOAD & TRANSFORM BILL XML DATA FROM LOCAL FILE**

def findheaderword(headers):
    verbspd=pd.DataFrame(columns=['BillID', 'HeaderID', 'HeaderOrderID','EntityOrderID', 'Text','PartofSpeech', 'SubjObjPred', 'HeadText']);
    billidlist=[];
    nlp = spacy.load("en_core_web_sm");
    for ind in headers.index:
    
        doc = nlp(headers.loc[ind,'Text'].replace('-', ' ').replace('     ', '')) ;
        span = doc[doc[len(doc)-1].left_edge.i : doc[len(doc)-1].right_edge.i+1];
        entorder=0 ;
        with doc.retokenize() as retokenizer:
            retokenizer.merge(span);
            for token in doc:
                entorder=entorder+1;
                verbspd=verbspd.append({ 'BillID': headers.loc[ind, 'BillID'], 'HeaderID': headers.loc[ind, 'HeaderID'],
                            'HeaderOrderID': headers.loc[ind, 'HeaderOrderID'],'EntityOrderID':entorder, 'Text':token.text,
                                           'PartofSpeech':token.pos_, 'SubjObjPred': token.dep_, 'HeadText':token.head.text},ignore_index=True);
        if str(headers.loc[ind, 'BillID'])[-1]=='0' and headers.loc[ind, 'BillID'] not in billidlist:
            billidlist.append(headers.loc[ind, 'BillID']);
            print(headers.loc[ind, 'BillID']);
    return verbspd

# **EXECUTE DATA IMPORT & TRANSFORM FUNCTIONS**

username='joythompson'
projfol="/Users/"+username+"/Desktop/DSE_203_JoyT/DSE_203_Final_Project_JoyT/"
directorybillhead=createfiledirectory(projfol+ 'Bill_Parsing_Headers/')
headerpd=pd.read_csv(directorybillhead+ 'header_text_by_bill_header_with_order.csv')

verbspd=findheaderword(headerpd)

verbspd.to_csv(directorybillhead+ 'bill_obj_action_subj_by_bill_header.csv')

LOAD CSV WITH HEADERS FROM "https://github.com/JoyT10294/DSE203_FinalProj/blob/main/polirepsdata.csv" AS repdata
CREATE (politician:Politician {id: repdata.representativeid, lastname: repdata.lastname, firstname: repdata.firstname})
CREATE (state:State {id: repdata.statecode, name: repdata.statelabel})

CREATE (politician)-[:REPRESENTS {district:repdata.district}]->(state)
CREATE (party:Party {name: repdata.party})
CREATE (politician)-[:MEMBER_OF]->(party)
CREATE (chamber:Chamber {name: repdata.chamber})
CREATE (politician)-[:SERVES_IN]->(chamber)

LOAD CSV WITH HEADERS FROM "https://github.com/JoyT10294/DSE203_FinalProj/blob/main/polirepsdata.csv" AS repdata
CREATE (politician:Politician {id: , lastname:, firstname: })
CREATE CONSTRAINT ON (politician:Politician) ASSERT politician.id IS UNIQUE
CREATE CONSTRAINT ON (politician:Politician) ASSERT exists(politician.id)
CREATE (state:State {id: , name:})
CREATE CONSTRAINT ON (state:State) ASSERT state.id IS UNIQUE
CREATE CONSTRAINT ON (state:State) ASSERT exists(state.id)
CREATE (politician)-[:REPRESENTS {district:repdata.district}]->(state)
CREATE (party:Party {name: repdata.party})
CREATE (politician)-[:MEMBER_OF]->(party)
CREATE (chamber:Chamber {name: repdata.chamber})
CREATE (politician)-[:SERVES_IN]->(chamber)
